move comm and ver keys to separate classes
