+++ circuits/cpp/barretenberg/cpp/src/barretenberg/honk/pcs/commitment_key_new.hpp
@@ -1,0 +1,83 @@
+#pragma once
+
+/**
+ * @brief Provides interfaces for different 'CommitmentKey' classes.
+ *
+ * TODO(#218)(Mara): This class should handle any modification to the SRS (e.g compute pippenger point table) to
+ * simplify the codebase.
+ */
+
+#include "barretenberg/ecc/curves/bn254/bn254.hpp"
+#include "barretenberg/ecc/curves/bn254/pairing.hpp"
+#include "barretenberg/ecc/curves/grumpkin/grumpkin.hpp"
+#include "barretenberg/ecc/scalar_multiplication/scalar_multiplication.hpp"
+#include "barretenberg/numeric/bitop/pow.hpp"
+#include "barretenberg/polynomials/polynomial.hpp"
+#include "barretenberg/polynomials/polynomial_arithmetic.hpp"
+#include "barretenberg/srs/factories/crs_factory.hpp"
+#include "barretenberg/srs/factories/file_crs_factory.hpp"
+
+#include <cstddef>
+#include <memory>
+#include <string_view>
+
+namespace proof_system::honk::pcs {
+
+// using Curve = curve::BN254;
+
+/**
+    * @brief CommitmentKey object over a pairing group 𝔾₁, using a structured reference string (SRS).
+    * The SRS is given as a list of 𝔾₁ points { [xʲ]₁ }ⱼ where 'x' is unknown. The SRS stored in the commitment key is
+    * after applying the pippenger_point_table thus being double the size of what is loaded from path.
+    *
+    *
+    */
+template <class Curve>
+class CommitmentKey {
+
+    using Fr = typename Curve::ScalarField;
+    using Commitment = typename Curve::AffineElement;
+
+    public:
+    CommitmentKey() = delete;
+
+    /**
+        * @brief Construct a new Kate Commitment Key object from existing SRS
+        *
+        * @param n
+        * @param path
+        *
+        */
+    CommitmentKey(const size_t num_points,
+                    std::shared_ptr<barretenberg::srs::factories::CrsFactory<Curve>> crs_factory)
+        : pippenger_runtime_state(num_points)
+        , srs(crs_factory->get_prover_crs(num_points))
+    {}
+
+    // Note: This constructor is used only by Plonk; For Honk the CommitmentKey is solely responsible for extracting
+    // the srs.
+    CommitmentKey(const size_t num_points,
+                    std::shared_ptr<barretenberg::srs::factories::ProverCrs<Curve>> prover_crs)
+        : pippenger_runtime_state(num_points)
+        , srs(prover_crs)
+    {}
+
+    /**
+        * @brief Uses the ProverSRS to create a commitment to p(X)
+        *
+        * @param polynomial a univariate polynomial p(X) = ∑ᵢ aᵢ⋅Xⁱ ()
+        * @return Commitment computed as C = [p(x)] = ∑ᵢ aᵢ⋅[xⁱ]₁ where x is the secret trapdoor
+        */
+    Commitment commit(std::span<const Fr> polynomial)
+    {
+        const size_t degree = polynomial.size();
+        ASSERT(degree <= srs->get_monomial_size());
+        return barretenberg::scalar_multiplication::pippenger_unsafe<Curve>(
+            const_cast<Fr*>(polynomial.data()), srs->get_monomial_points(), degree, pippenger_runtime_state);
+    };
+
+    barretenberg::scalar_multiplication::pippenger_runtime_state<Curve> pippenger_runtime_state;
+    std::shared_ptr<barretenberg::srs::factories::ProverCrs<Curve>> srs;
+};
+
+} // namespace proof_system::honk::pcs
